+++ app/Models/UserLedgerEntry.php
@@ -20,10 +20,13 @@
 
 namespace Adshares\Adserver\Models;
 
+use Adshares\Adserver\Facades\DB;
 use Adshares\Common\Exception\InvalidArgumentException;
+use DateTime;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use Illuminate\Database\Query\Builder as DatabaseBuilder;
 use Illuminate\Support\Facades\Log;
 use function array_filter;
 use function array_merge;
@@ -35,6 +38,7 @@ use const PHP_INT_MAX;
 /**
  * @mixin Builder
  * @property int status
+ * @property int type
  */
 class UserLedgerEntry extends Model
 {
@@ -129,6 +133,7 @@ class UserLedgerEntry extends Model
         'id' => 'int',
         'amount' => 'int',
         'status' => 'int',
+        'type' => 'int',
         'user_id' => 'int',
     ];
 
@@ -434,5 +439,65 @@ class UserLedgerEntry extends Model
         }
 
         return $type;
+    }
+
+    public static function getBillingHistoryBuilder(
+        int $userId,
+        array $types,
+        ?DateTime $from,
+        ?DateTime $to
+    ): DatabaseBuilder {
+        return DB::table('user_ledger_entries')->select(
+            [
+                DB::raw('MAX(id) AS id'),
+                DB::raw('SUM(amount) AS amount'),
+                'type',
+                'status',
+                'address_from',
+                'address_to',
+                'txid',
+                DB::raw('MAX(created_at) AS created_at'),
+            ]
+        )->fromSub(
+            function (DatabaseBuilder $subQuery) use ($userId, $types, $from, $to) {
+                $subQuery->from('user_ledger_entries')->select(
+                    [
+                        'id',
+                        'amount',
+                        'type',
+                        'status',
+                        'address_from',
+                        'address_to',
+                        DB::raw(
+                            DB::isSQLite()
+                                ? 'CASE WHEN type IN (3, 4, 5, 6) AND status = 0 THEN null ELSE txid END AS txid'
+                                : 'IF(type IN (3, 4, 5, 6) AND status = 0, null, txid) AS txid'
+                        ),
+                        'created_at',
+                        DB::raw(
+                            DB::isSQLite()
+                                ? 'CASE WHEN type IN (3, 4, 5, 6) AND status = 0 '
+                                .'THEN date(created_at) ELSE created_at END AS date_helper'
+                                : 'IF(type IN (3, 4, 5, 6) AND status = 0, date(created_at), created_at) AS date_helper'
+                        ),
+                    ]
+                )->where('user_id', $userId);
+
+                if (!empty($types)) {
+                    $subQuery->whereIn('type', $types);
+                }
+
+                if (null !== $from) {
+                    $subQuery->where('created_at', '>=', $from);
+                }
+
+                if (null !== $to) {
+                    $subQuery->where('created_at', '<=', $to);
+                }
+
+                return $subQuery;
+            },
+            'u'
+        )->groupBy(['date_helper', 'type', 'status', 'txid', 'address_from', 'address_to']);
     }
 }