+++ deploy/crontablist.sh
@@ -32,6 +32,10 @@ echo -n " && "
 echo -n "php ${SERVICE_DIR}/artisan ops:adpay:event:export ${_REDIRECTION}"
 echo ""
 
+echo -n "* * * * * "
+echo -n "php ${SERVICE_DIR}/artisan ops:expired-withdrawal:cancel ${_REDIRECTION}"
+echo ""
+
 echo -n "30 * * * * "
 echo -n "php ${SERVICE_DIR}/artisan ops:adpay:payments:get"
 echo -n " && "
@@ -62,6 +66,10 @@ echo ""
 
 echo -n "0 0 * * * "
 echo -n "php ${SERVICE_DIR}/artisan ops:license:fetch"
+echo ""
+
+echo -n "0 0 * * * "
+echo -n "php ${SERVICE_DIR}/artisan ops:serve-domains:update"
 echo ""
 
 test ${SKIP_COLD_WALLET:-0} -eq 0 && \