+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.2.0] - 2019-09-11
+### Added
+- Cancel expired withdrawal automatically
+- Billing history filtering
 ### Changed
 - Import inventory: reject unclassified banners
 
@@ -196,7 +201,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v1.1.2...develop
+[Unreleased]: https://github.com/adshares/adserver/compare/v1.2.0...develop
+[1.2.0]: https://github.com/adshares/adserver/compare/v1.1.2...v1.2.0
 [1.1.2]: https://github.com/adshares/adserver/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/adshares/adserver/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/adshares/adserver/compare/v0.12.0...v1.1.0